Add new blog posts and update website layout
Added five new blog posts and created a new Blog section in the website navigation. Adjustments to the main layout include default title and description values and have the slot inside a main HTML tag. Additionally, added 404, Contact, and About pages. The CSS was overridden to modify the style of headers and paragraphs and add styles specific to 404 page.